Начинаем работать над апишкой

+++ shop_server/models/product.go
@@ -1,8 +1,8 @@
 package models
 
 type Product struct {
-	ID         int    `json:"id"`
-	Name       string `json:"name"`
-	Price      int    `json:"price"`
-	Categories string `json:"categories"`
+	ID         int     `json:"id"`
+	Name       string  `json:"name"`
+	Price      float64 `json:"price"`
+	Categories string  `json:"categories"`
 }
